- resolve #245: Refactor Common and Utils namespaces types. - resolve #246: Remove unused namespaces across the codebase.
