+++ Rollbar/DTOs/TelemetryLevel.cs
@@ -2,10 +2,7 @@
 {
     using Newtonsoft.Json;
     using Newtonsoft.Json.Converters;
-    using System;
-    using System.Collections.Generic;
     using System.Runtime.Serialization;
-    using System.Text;
 
     /// <summary>
     /// Enumerates supported telemetry levels.
