+++ Rollbar/Deploys/RollbarDeployClient.cs
@@ -48,7 +48,7 @@
                 return this._httpClient;
             }
 
-            return HttpClientUtil.CreateHttpClient(this._config.ProxyAddress, this._config.ProxyUsername, this._config.ProxyPassword);
+            return HttpClientUtility.CreateHttpClient(this._config.ProxyAddress, this._config.ProxyUsername, this._config.ProxyPassword);
         }
 
         private bool Release(HttpClient httpClient)
