+++ Rollbar/Instrumentation/InstrumentationCondition.cs
@@ -1,9 +1,6 @@
 ﻿namespace Rollbar.Instrumentation
 {
     using System;
-    using System.Collections.Generic;
-    using System.Diagnostics;
-    using System.Text;
 
     /// <summary>
     /// Class InstrumentationCondition.
