+++ Rollbar/Telemetry/TelemetryCollector.cs
@@ -4,7 +4,6 @@ namespace Rollbar.Telemetry
 {
     using System;
     using System.Diagnostics;
-    using System.Runtime.InteropServices;
     using System.Threading;
     using Rollbar.DTOs;
